refactor(tests): reorganize revert tests by EIP and improve assertions
- Move CREATE2 revert tests to tests/constantinople/eip1014_create2/
- Move warm/cold status revert tests to tests/berlin/eip2929_gas_cost_increases/
- Remove duplicate TSTORE revert test (already covered by eip1153 tests)
- Keep only SELFDESTRUCT balance transfer revert in eip6780 directory
- Use storage.canary() to verify zero values are actually written
- Minimize all non-zero balances to 1
- Add deployer post-state assertion in collision test
- Add beneficiary post-state assertion in collision test
